+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 # 0.2.0
 
-Upgraded to bevy 0.17
+* Upgraded to bevy 0.17
+* Added additional capabilities and renamed some in `bevy_immediate::ui`
+* Added support for bevy_feathers capabilities. Use set `CapsUiFeathers`
+* Added support for couple of bevy_ui_widgets features. Use set `CapsUiWidget`
 
 # 0.1.1